+++ acme2certifier/django_app/views.py
@@ -20,6 +20,7 @@ from acme2certifier.acme_srv.helper import (
     config_check,
     legacy_acme_get_load,
     acme_get_method_not_allowed_problem,
+    server_name_configuration_validate,
 )
 from acme2certifier.acme_srv.db_handler import log_active_db_handler
 from acme2certifier.acme_srv.housekeeping import (
@@ -82,6 +83,7 @@ ERR_RESPONSE_ACME_GET["Allow"] = "POST"
 
 # check configuration for parameters masked in ""
 config_check(LOGGER, CONFIG)
+server_name_configuration_validate(LOGGER, CONFIG)
 LEGACY_ACME_GET = legacy_acme_get_load(LOGGER, CONFIG)
 
 with Housekeeping(DEBUG, LOGGER) as housekeeping: